db2_ldap: Rename module interface file #82
